CAN is sending correctly now

+++ Core/Inc/corner_driver.h
@@ -4,4 +4,4 @@
 #include "stm32f4xx_hal_gpio.h"
 
 void Read_ADC_Data(SPI_HandleTypeDef *hspi, uint8_t *spi_rx);
-void Read_Internal_ADC_Data(ADC_HandleTypeDef *hadc, uint16_t *adc_val);
+void Read_Internal_ADC_Data(ADC_HandleTypeDef *hadc, uint32_t *adc_val);
