+++ Core/Src/corner_driver.c
@@ -10,10 +10,10 @@ void Read_ADC_Data(SPI_HandleTypeDef *hspi, uint8_t *spi_rx) {
   // }
 }
 
-void Read_Internal_ADC_Data(ADC_HandleTypeDef *hadc, uint16_t *adc_val) {
+void Read_Internal_ADC_Data(ADC_HandleTypeDef *hadc, uint32_t *adc_val) {
   HAL_ADC_Start(hadc);
   if (HAL_ADC_PollForConversion(hadc, 100) == HAL_OK) {
-    *adc_val = (uint16_t)HAL_ADC_GetValue(hadc);
+    *adc_val = HAL_ADC_GetValue(hadc);
     // printf("Internal ADC Value: %lu\n", *adc_val);
   } else {
     printf("Failed to read internal ADC value.\n");
